Change End Soc to equal start SOC for an IOG hold slot
The .flows calculation is EV unaware, so assumes the inverter is in self use for EV charging slots. Thus End SOC is reported as less than Start SOC in the Dashboard for an IOG slot, which is incorrect as the inverter will be holding SOC. 

To fix this, for an IOG slot, make 'windows' soc_end equal to soc.